compose 使わない activity を追加
